Anne : getCategorieById

+++ ProjetEcommerce/src/fr/adaming/dao/ICategorieDao.java
@@ -9,7 +9,7 @@ import fr.adaming.model.Categorie;
 @Local
 public interface ICategorieDao {
 	public List<Categorie> getAllCategorie();
-	public Categorie getCategorieById(int id);
+	public Categorie getCategorieById(long id);
 	public Categorie addCategorie(Categorie c);
 	public int deleteCategorie(Categorie c);
 	public int modifyCategorie(Categorie c);
